+++ docs/source/changelog.rst
@@ -3816,5 +3816,9 @@ Current C++/SQLite client, Python/SQLAlchemy server
   the patient.
   https://github.com/ucam-department-of-psychiatry/camcops/issues/263
 
+- Support for a clinician to configure IDED-3D settings for single user mode on
+  a per-patient basis.
+  https://github.com/ucam-department-of-psychiatry/camcops/issues/314
+
 - Supported SQLAlchemy version now 1.4
   https://github.com/ucam-department-of-psychiatry/camcops/issues/172